+++ lib/virtio/virtqueue.c
@@ -359,9 +359,9 @@ void *virtqueue_get_available_buffer(struct virtqueue *vq, uint16_t *avail_idx, 
 
     env_rmb();
 #if defined(RL_USE_ENVIRONMENT_CONTEXT) && (RL_USE_ENVIRONMENT_CONTEXT == 1)
-    buffer = env_map_patova(vq->env, ((uint32_t)(vq->vq_ring.desc[*avail_idx].addr));
+    buffer = env_map_patova(vq->env, ((uintptr_t)(vq->vq_ring.desc[*avail_idx].addr)));
 #else
-    buffer = env_map_patova((uint32_t)(vq->vq_ring.desc[*avail_idx].addr));
+    buffer = env_map_patova((uintptr_t)(vq->vq_ring.desc[*avail_idx].addr));
 #endif
     *len = vq->vq_ring.desc[*avail_idx].len;
